dim-web: initial commit
