+++ dim-utils/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "dim-utils"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+dia-i18n = "0.10.0"
